# labels/cidr: Fix slice preallocation size